northd: Fixed stage name mapping in lflow table.
When changing the order of stages in the LogicalFlow
table in SB, some rules will be added with an incorrect
stage-name in the external-ids of existing datapath groups.
When adding a stage, new lflows in a new table with the same
match and action that already existed with the same table-id
will remain in the database with unupdated stage names.

Signed-off-by: Alexandra Rukomoinikova <arukomoinikova@k2.cloud>