RFC: Adding a new TLV for transferring the route table.
The patch below is very messy and only a proof of concept. Its further development is
impossible without solving the issue with registers, which I described in the cover letter.

Problem that the patch solves:
Currently, for each routing table we need to create a separate switch,
because the routing table is determined by the port from which the traffic arrived.

As shown in the diagram below, for cross-AZ deployment.
      +---------------+                         +---------------+
      |  subnet 1     |                         |  subnet 2     |
      | route table 1 |                         | route table 2 |
      +-------+-------+                         +-------+-------+
              |                                       |
              |               +---------------+       |
              +-------------->|    Router     |<------+
                              +-------+-------+
                                      |
                    +-----------------+----------+
                    |                            |
              +-----+--------+            +------+-------+
              | TS Router    |            | TS Router    |
              |  Table 1     |            |   Table 2    |
              +-----+-------+              +------+-------+
                                   AZ2

After these changes:
The route selector will store the route table and send it to the tunnel in the
GENEVE header option field.
Based on this information, when processing on the router in the second availability zone,
the correct routing table will be selected, and a separate switch for each route
table will no longer be needed.
In the future, this functionality will be hidden behind an optional flag
to avoid breaking offloading in regular deployments.

Signed-off-by: Alexandra Rukomoinikova <arukomoinikova@k2.cloud>